port IndentingLexerCasesDedentTest to zio-test
